+++ src/domain.rs
@@ -3,7 +3,6 @@ pub mod constraints;
 pub mod field_vector;
 pub mod grid;
 pub mod mass_history;
-pub mod momentum_residual;
 pub mod parity;
 pub mod simulation_output;
 pub mod state;
